requirements.txt
PySide6==6.6.0
torch==2.10.0
python-dotenv==1.0.0
requests==2.32.5
langchain-core==1.2.17
pydantic==2.12.5
langchain-community==0.4.1
openai==2.26.0
sounddevice==0.4.6
numpy==1.26.0
pydub==0.25.1